authentication_backend/requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-cors-headers==4.0.0
django-formtools==2.4.1
django-otp==1.2.1
django-phonenumber-field==6.4.0
django-two-factor-auth==1.15.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
phonenumbers==8.13.13
PyJWT==2.7.0
pypng==0.20220715.0
pytz==2023.3
qrcode==7.4.2
requests==2.31.0
sqlparse==0.4.4
twilio==8.2.1
typing_extensions==4.6.2
urllib3==2.0.2
yarl==1.9.2
